services/wycena/requirements.txt
boto3~=1.34.11
fastapi~=0.114.0
googlemaps==4.10.0
openpyxl==3.1.2
pydantic==2.9.1
pydantic-settings==2.4.0
pynamodb~=5.5.1
pytest~=7.4.4
python-multipart
sqlmodel==0.0.22
strawberry-graphql[fastapi]
SQLAlchemy~=2.0.25
